Move all lib code into src folder.
