+++ lib/src/configs/default_settings.dart
