+++ lib/src/enums/supported_input_file_type.dart
