+++ lib/src/services/flappy_logger.dart
